Register input types for model fields
Instead of having the input types created when the fields for a mutation are
retrieved, a method is defined to fetch the appropriate input type depending
on the provided model field. This enables a flexible method to register
new input types for custom model fields and even allows to override input types
for existing model fields.